fix: unable to switch API version on singleton
